+++ Changes.md
@@ -16,6 +16,8 @@ Improvements
     - Organised colorspaces into submenus by family.
     - Removed unwanted title-casing, so that names are now displayed verbatim.
     - Added control over the presence of roles by registering `openColorIO:includeRoles` metadata to the relevant plugs.
+  - Improved View display transform menu :
+    - Ordered menu items to match the order in the OpenColorIO config.
 - VectorDataPlugValueWidget : Computation errors are now reflected by a red background colour.
 
 Fixes